+++ NuKeeper.Inspection.Tests/Report/CsvReportFormatTests.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using NuGet.Packaging.Core;
 using NuGet.Versioning;
 using NuKeeper.Inspection.Report;
+using NuKeeper.Inspection.Report.Formats;
 using NuKeeper.Inspection.RepositoryInspection;
 using NUnit.Framework;
