+++ NuKeeper.Inspection/Report/Formats/NullReportFormat.cs
@@ -1,27 +1,20 @@
 using System.Collections.Generic;
 using NuKeeper.Inspection.RepositoryInspection;
 
-namespace NuKeeper.Inspection.Report
+namespace NuKeeper.Inspection.Report.Formats
 {
-    public class LibYearsReportFormat : IReportFormat
+    public class NullReportFormat : IReportFormat
     {
         private readonly IReportWriter _writer;
 
-        public LibYearsReportFormat(IReportWriter writer)
+        public NullReportFormat(IReportWriter writer)
         {
             _writer = writer;
         }
 
         public void Write(string name, IReadOnlyCollection<PackageUpdateSet> updates)
         {
-            var totalAge = Age.Sum(updates);
-            var years = totalAge.TotalDays / 365;
-
-            _writer.WriteLine(years.ToString("0.0"));
-
             _writer.Close();
         }
-
-
     }
 }
